modify file db 상태 변경 및 form으로 raw_len 받아오면 적용시키는 걸로 변경 만약 데이터를 못받아오면 print 됨. default 값은 60*60(1시간) 으로 설정
